[Autofill Assistant] Refactor |SetFieldValue|
Next step of refactoring |WebController|. Extract |SetFieldValue| and as
a dependant also |SendKeyboardInput| into the util. This will allow the
next step to extract the chain used in the |ClickOrTap| to be moved
to the util and out of the |WebController|.

From there it's up to the committee to decide how to continue. The chain
should not remain hardcoded but somehow be moved into the action.

Bug: b/158153191
Change-Id: If2f23e4423f88757a172fef2fd76d6884e1408dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255362
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Luca Hunkeler <hluca@google.com>
Cr-Commit-Position: refs/heads/master@{#784752}