policy-testserver: Added functionality to issue certificates
Added the functionality to reply to client_cert_provisioning requests
required for certain policies, i.e. RequiredClientCertificateForDevice.
A certificate is generated based on the public key recieved from the
requestor and a hard coded private key. The certificate is then sent as a
reply to the requestor.

BUG=None
TEST=tast run <ip> policy.RequiredClientCertificateForDevice

Change-Id: I88f6bb0b7a0c06cb4984da36e35e29cf2e84b806
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267159
Commit-Queue: Alexander Hartl <alexanderhartl@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Michael Ershov <miersh@google.com>
Cr-Commit-Position: refs/heads/master@{#784782}