bookmarks: Enforce the construction of valid permanent nodes
This patch introduces factory functions that enforce how
BookmarkPermanentNode instances are created, with consistent values,
most importantly the GUID, which had a subtle implementation prior to
this patch.

The previously-existing function PermanentNodeTypeToGuid() was flawed
when it came to the managed bookmark node, since there is no explicit
value in enum BookmarkNode::Type to identify managed bookmarks, and
FOLDER was used to conclude that it must be the managed bookmarks
folder.

Instead, it's cleaner and less fragile to force all callers of
BookmarkPermanentNode's constructor to specify a GUID.

Change-Id: Ife11f5289053fdcb99b1a65d38accf9d80c9e89e
Bug: 1060311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083479
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784812}