cert_provisioning: Display failed processes as Failed in the UI
The UI was mistakenly displayed FailedWorkerInfo::state in the UI, which
is the state prior to failure, so it will never say Failed.

Clarify the meaning of the field by renaming/comments and always display
failed for failed workers.

In a follow-up, the status text will be changed to Failed (was <other
status>) or similar. This is not done in this CL in case we want to
merge it back, which does not allow UI string changes.

Bug: 1098244
Change-Id: Ib7c2a8308211bb4459640a99dd44e7c5c06dc23f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276420
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Michael Ershov <miersh@google.com>
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784836}