Temporary CHECK that default factory always has an opaque initiator.
This CL adds a temporary CHECK to verify in the wild the hypothesis
that the default URLLoaderFactory is only ever used for requests
where network::ResourceRequest::request_initiator is an opaque origin.
This CL should be reverted once we gather enough data to confirm
or disprove this hypothesis / assertion.

See also the discussion in network-service-dev@chromium.org at
https://groups.google.com/a/chromium.org/d/topic/network-service-dev/hCk6fJJkNF4/discussion

Change-Id: I846937c04583164880cb5893a350d26ec30b0048
Bug: 1098938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276756
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784877}