Null check in ThroughputUkmReporter::ReportThroughputUkm
It looks like in this function, the ukm_manager_ could be null,
and that could cause crash. I think it is possible being null
when a tab is shutting down and that the ukm_manger_ has
been cleared.

This CL does null check and thus prevent crash.

Bug: 1100868
Change-Id: I6efd5064eed1aa60cf45d087677bdb49d08e70b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277348
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784920}