Add test for RenderFrameHostImpl::Reload()
This function was not tested, that's unfortunate. It currently
corresponds to an IPC send toward the renderer process. The renderer
process then "tries" to reload the document, by requesting the browser
process to navigate. This is a useless process round-trip. Moreover, the
renderer doesn't know everything. Some useful information about the
history is centralized inside the browser process. The renderer process
reload the document imperfectly. Several bit of information are
missing.

This patch adds a test, checking how much information is lost.
Surprisingly, this is not too bad. Only the initiator is incorrect.

horo@ is going to provide a real good implementation, based on the
browser process:
https://chromium-review.googlesource.com/c/chromium/src/+/2275567

Thanks to this test, we are going to see how much we recover.

Bug: chromium:995428
Change-Id: I183a0a401e2b067259f282d27aabc90d8a0f9b19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279774
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784996}