[signin] Customize pre-reauth dialog buttons depending on reauth type
This CL modifies the reauth confirmation dialog to hide action buttons
until it's known whether the reauth is auto-approved. The dialog
shows a spinner while the buttons are hidden.

A screencast of the result flow is available in the attached bug.

To let the reauth controller trigger an event in webUI,
SigninReauthHandler starts depending on SigninReauthViewController*
instead of Browser*. This is fine since SigninReauthViewController is
guaranteed to outlive SigninReauthHandler. SigninReauthViewController
owns a native view hosting the webUI.

Bug: 1100923
Change-Id: I6bcd2a31e93379ee60497be4eae75cd712a8a1d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274628
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785012}