ExternalVkImageBacking: only do synchronization when GL texture is created
For some use cases, the ExternalVkImageBacking is created with
SHARED_IMAGE_USAGE_GLES2, but it is not used by GL (WebGL) at all.
For example:video playback. On linux, video is decoded into
ExternalVkImageBackings with SHM. So decoder fills the SHM,
and then pixels are uploaded form SHM to VkImage, and then it will
be consumed by compositor for rendering frame. In most cases, it will
never be used as GL texture (for WebGL). So doing synchronization
is costly and unnecessary. Address this performance issue by only
doing synchronization when ProduceGL*() is called.

Bug: 1087063
Change-Id: I09eff667fa393f3c3a26da639130ad385797f5f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279927
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Peng Huang <penghuang@chromium.org>
Auto-Submit: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785095}