[IntentHandling] Link files app to query task from app service.
This CL creates an app service file tasks to allow the files app to
query sharing task from app service intent handling backend. This is for
testing the backend and it is behind a flag. It might need to be
modified after link up with unified sharesheet UI.

BUG=1092784

Change-Id: I4ba54f7b2ee9e9ae3fd89184b808c1040a01c288
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279411
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785316}