vulkan: fix a crash in SkiaOutputDeviceVulkan.
The crash is because we assume vulkan driver will always create
swapchain with the number of images which is specified by
VkSwapchainCreateInfoKHR::minImageCount, but sometimes, the driver
will not obey it, and then it causes problems. fix the crash by using
the number from the swapchain instead of assuming it.

Bug: 1099132
Change-Id: I70868d1bf9efee4d845ff21a236af33eac401ad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281177
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Auto-Submit: Peng Huang <penghuang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785357}