[Autofill Assistant] Remove execution context from generic interactions.
I introduced the context originally to avoid having to replace
placeholders without reflection. However, this turned out to become
intractable, as the number of places where we'd need to account for
placeholders increased. Since the upstream of this CL adds the relevant
placeholder logic, we can now do the replacement directly in the proto,
rather than at runtime when looking up views or values.

Bug: b/145043394
Change-Id: I384dd3828183c4f616db15bf3a5d202eaf04ea01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274627
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#785427}