base/allocator: Handle aligned allocations outside of GigaCage.
Aligned allocations are incompatible with adding the generation of the
allocation before the content. A solution is to exclude aligned
allocations from the tagged pointers, that is to say make sure they are
allocated from outside the GigaCage.

This approach is used in this patch, by:
- Adding a |use_giga_cage_if_available| in PartitionRoot
- Creating a new PartitionRoot in the allocator shim, handling aligned
  allocations.
- Updating PartitionBucket and PartitionPage to not use the GigaCage
  when the PartitionRoot doesn't support it.
- Making PartiionRoot::Free() static.

The last point is important: As posix_memalign() is not matched by a
special free() function. This means that PartitionRoot::Free() can be
called with a pointer from a different root (namely the aligned or
unaligned partitions). Fortunately, the root can always be found from
the pointer address, through PartitionPage which we need anyway. This
costs an additional masking operation and a pointer indirection, though
this also removes an atomic load.

Co-authored with bartekn@chromium.org.

Bug: 1092288, 998048
Contributor: bartekn@
Change-Id: Ib3cdd70c7e090d0e90ca7b4edd8e0410e51a4829
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275886
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785458}