Fix the ChromeVox panel.
This patch is similar to the following: http://crrev.com/c/2055415

The issue here is that the LocaleOutputHelper class never gets
initialized in the panel context. The fix is to call its init() method
when initializing the panel.

The reason we didn't catch this earlier is because we didn't have a
panel test that ran with language switching enabled. This patch adds
that test so we can be certain this issue won't occur in the future.

Bug: 1101057
AX-Relnotes: N/A
Change-Id: I389caf9e2dcebd5113436349b27bb54c5911bfb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274839
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785471}