Add LockScope to ScopedSafearray for random access and iteration
Adding a templatized class LockScope to ScopedSafearray which allows
random access and iteration over the values stored in a SAFEARRAY for
any VARTYPE that is implemented by VariantUtil<VARTYPE>.

To create the lock scope, ScopedSafearray::CreateLockScope attempts
to lock the underlying SAFEARRAY. If VARTYPE is expected, and the lock
succeeded, returns a LockScope.

When LockScope is destructed, it automatically releases the lock
granted by ScopedSafearray.

Also adding ScopedSafearray::GetCount to get the number of elements
stored in a safearray for a given dimension.

Split out from :
https://chromium-review.googlesource.com/c/chromium/src/+/2188057

Bug: 928948
Change-Id: Ibbe79385ecdabcae5801c867321560bd37c7dbe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244457
Commit-Queue: Adam Ettenberger <Adam.Ettenberger@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#785477}