ozone/wayland: Fix capability event handling
The current capability handling code does not cleanly handle cases
where the new capability mask contains flags that were already set
in the old capability mask.

In particular, if both the old and the new mask contain the pointer
capability, a new WaylandPointer will be constructed before the
old one is destroyed, causing WaylandEventSource to receive a
OnPointerCreated event for the new pointer before receiving an
OnPointerDestroyed event for the old pointer.

In release builds, the way this manifests is that the
OnPointerDestroyed handler clears out the new pointer that was just
assigned in the OnPointerCreated handler. (Debug builds have a
DCHECK_EQ that will abort execution instead.)

Interestingly, this used to be handled correctly; but commit
1f7cb5d6ed66 ("ozone/wayland: Factor EventSource out of Connection
class"), which seems to have been intended as pure refactoring,
broke it. Restore the behavior from before that commit.

Change-Id: I0f7c677e0bb9ae016112c7224da78d3201c27020
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279889
Commit-Queue: Jann Horn <jannh@google.com>
Auto-Submit: Jann Horn <jannh@google.com>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#785485}