Prepare RenderTests for internal support
Prepares the Android RenderTests for upcoming support for running on
internal bots and keeping the results separate from the public results.

Changes are:
1. Require that a corpus be specified and define an internal corpus.
2. Change all uses of RenderTestRule to use the builder and specify the
   public corpus.
3. Include model/SDK version in the JSON uploaded to Gold.
4. Rename RenderTestRule.SkiaGoldBuilder to RenderTestRule.Builder since
   the distinction is no longer necessary.

There should be no functional difference in the tests with this CL, but
will ensure that results that are meant to be public are forwarded to
the public Gold instance (and internal results are kept private) once
mirroring is set up.

Bug: 1077274
Change-Id: I37fca15875ed820f18cf8615bbe50e77efe6ac74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274365
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785523}