[scheduling] Experiment to prioritize before unload handlers.
This CL is the first step in a series of experiments around prioritizing the unload pathway in the scheduler. The design doc[1] has more details.

This CL adds a new mojo interface 'UserBlockingLocalFrame' which allows the browser to send high priority messages to a local frame. These messages will run at a higher priority than legacy IPC messages, but lower than input. We will use this channel to send calls to LocalFrame::BeforeUnload so that they will be prioritized in the renderer.

All changes are implemented behind a flag so we can experiment with finch. This CL also contains some browser test fixes around timing that popped up when the feature was enabled by default. Also note that this only impacts browser-initiated navigations since renderer-initiated navigations dispatch the before unloads synchronously with the script that triggers the navigation, already effectively running as soon as they can.

[1] https://docs.google.com/document/d/1srzGhb1eCS66mCl2y8CqjN2o3iTPXI2WfySIKoXZQKo/edit?usp=sharing

Change-Id: Ia131bb65087f2612a1dd162f11cafe53862d787b
Bug: 1042118
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173637
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785556}