base: Fix bindService with empty intent
This is a follow up bug fix from
https://chromium-review.googlesource.com/c/chromium/src/+/2233312

Bug happens when:
The fallback timeout is triggered at once so sFallbackEnabled
Try to bind to a non-sandboxed process, which does not have a fallback
so mFallbackServiceName is null.

In this case, the intent sent to bindService is just empty, thus it's
going to fail. Fix by checking for mFallbackServiceName being null, and
avoid using it if that's the case.

Bug: 994961
Change-Id: I51e62898c0c1cadd4a89ce184c59ba47e2b1f5c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283393
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785567}