Add device management (DM) request/response parser.
This is part of the changes for enterprise support in 
Chromium updater. The added code is to construct the 
request send to DM server and parse the response into 
an internal data structure.

This feature is a rewrite of a similar feature already
present in Omaha (Google Update).

This CL also introduces the dependency on //crypto. This is to enable
the parser to verify the SHA256 signature in the response and to
extract the public key for future response verification. In the unit
test, crypto lib is also used for signature creation. 

Bug: 1068797
Change-Id: Ieaa3565ff7293cfb52858ae39d18a794ef7a8823
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276535
Commit-Queue: Xiaoling Bao <xiaolingbao@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785582}