GestureNav: Fix a bug removing tab history popup immediately
This CL fixes a bug that removes tab history popup immediately when
releasing a long press. Now that the tab history is the only client
using NavigationSheet, this check gets easier by seeing if there's
already a navigation sheet running, and cancelling the key-up event in
that case. Android version check was removed since this is reported to
happen on later versions as well.

Bug: 1023018
Change-Id: I0885291d7b91c1205904efded3b73a8fb5ec5337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279256
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785625}