[WebLayer] Implement manage accounts API in WebLayer
This API uses the Mirror protocol to intercept any requested GAIA
account changes, and forwards them to a callback. The core logic is
implemented in weblayer::SigninURLLoaderThrottle and is similar to
chrome's signin::URLLoaderThrottle.

One difference in behavior between WebLayer and Chrome's handling of
Mirror, is that WebLayer skips the "Choose an account" page when a
GoogleAccountsCallback is set. This is done by adding a manage=true
query parameter to the /SignOutOptions page. This matches the behavior
we want for WebLayer and was suggested by the Google sign in team.

Bug: 1054160
Change-Id: Ibb316321105418d12901e91cf2a0b569f65e3bdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277337
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785627}