Rename files in c/b/extensions to remove blacklist.
This is the preparing for CL:2279412, which will replace blacklist for
the directory c/b/extensions.

The content of files will be changed in CL:2279412, and not changed in
this CL, because if the content is change, then gerrit marks it as
removed and added, which could be hard to view the change. So in this
CL, only file names are changed. Once this CL is submitted, I will
submit CL:2279412 for review to replace all blacklist for the
directory c/b/extensions.

FYI: https://tools.ietf.org/id/draft-knodel-terminology-00.html

Bug: 1097644
Change-Id: Ib973437600d4dbb3cd61597c3e071a9121d0ce8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279306
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785654}