Remove the flag kAppServiceContextMenu in c/b/ui/ash/launcher.
This is the preparing for removing the flag kAppServiceContextMenu.

The AppService context menu feature has been enabled for a few release,
and the flag is expired, so it can be removed in M86 release.

There will be a few follow up CLs to remove the flag from other parts of
the code.

BUG=1016159

Change-Id: I52c549e2300a579f0059f25ae35bb808e81cb64e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279261
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785661}