Lacros: Handle termination by Mojo disconnection.
With this CL, LacrosManager starts using Mojo disconnection
as a signal of lacros-chrome termination.
There can be a scenario that lacros-chrome is still alive
but Mojo error happens. However, even in the case, because
there's no way for ash-chrome to communicate with lacros-chrome,
lacros-chrome needs to be terminated anyways.
To handle async operations, introduces State.

Bug: 1091863
Test: Run locally in a device.
Change-Id: Ic566dbaeaeed387d931a737a2b4946d020cd82c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276411
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785667}