Validate index in page of page objects sent to PdfAccessibilityTree
Page object data that PdfAccessibilityTree receives have |index_in_page|
property. It is basically the index on that page object in the list of
that type of page object. For e.g. |page_in_index| of a link is the
index of the link in vector of links.

The validation implemented in this change is to check that
|index_in_page| is with in bounds of the page object vector.

Change-Id: Ic3ccf0aa2d988adbe1f6e8dd373b0852c96cb4db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279250
Commit-Queue: Virender Singh <virens@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ankit Kumar 🌪️ <ankk@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#785674}