Stop caching LabelButton's preferred size.
Caching was introduced in 2014, when computing the preferred size
involved repeatedly changing the text on the member label, laying it
out, and computing a bunch of text metrics.  All of that is gone now,
and computing the preferred size should be reasonably performant; it
can be made even more performant in the future by switching to a proper
layout manager.

In the meantime, this caching code not only complicates the code, it
makes it fragile; the existing implementation has places that reset the
cached value without actually firing PreferredSizeChanged(), for
example, which looks bogus.

Bug: none
Change-Id: I9cda6ba526d9fceda7a86b07a1475a844c6cbae7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283897
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785678}