SQLite: Add WAL mode support to CorruptSizeInHeader
CorruptSizeInHeader is a test helper function that allows to simulate
a corrupt database by altering the header of the DB file. In WAL mode,
transactions don't always touch the DB file, which means the database
can behave normally even when corrupted in this way. In this change,
the CorruptSizeInHeader method checkpoints and truncates the WAL
file. This ensures that any subsequent transaction is guaranteed to hit
the DB file and hence be affect by the header corruption. This change
also adds a test that verifies that the function works correctly.

Change-Id: Ic584e29f89b09a886d96e4424a17de8d37683314
Bug: 78507
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278413
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785702}