[PwdCheckAndroid] Add model and mediator for compromised credentials
This CL adds the (still unused) model and controller for the password
check feature. As sample entry, it uses a method to update credentials
and transforms it into elements that can be bound to items in a list
view. Neither data nor names are final but the basic data structure
(i.e. a ListModel) is strictly required to bind views and test them.

Binding the (already existing) view with a view binder and a MCP happens
in a follow-up.

READMEs are stealing heavily from touch_to_fill since their general
structure aligns fairly well.

Bug: 1101256,1092444
Change-Id: Ic85a8c36b01ee89e617706372e512cb7f5148119
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275959
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785704}