Clipboard API: Move encoding to the background thread.
In the current clipboard reader implementation, all of the encoding
was done on the main thread. This CL moves all of the encoding from
the image and text reader onto the background thread. The main benefit
is avoiding main thread jank when the user pastes a large payload from
the clipboard.

Bug: 1091833
Change-Id: I640f5818cfd81312f3adb59f4643c8e4dd34dc8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233982
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785738}