[CBD] Wire password deletion failure to the UI
Before this CL, BrowsingDataRemover did not expose the concept of
"deletion failure", i.e. when a requested deletion could not be
completed. (An example of such a deletion failure is a timeout, which
BrowsingDataRemoverImpl sets to 30 seconds.)

This CL adds the concept of "failed data types" to the API. They're
represented as a bitmask of DataTypes - the same format that's used to
request deletion in the various Remove*() methods.

So far, the only data type that will actually report deletion errors is
DATA_TYPE_PASSWORDS, but support for more types can easily be added.
None of the clients/observers does anything with the new information
yet; that's left for a future CL.

Bug: 1086433
Change-Id: I24b62c4da7d5c2739f47f3b599918d4643758411
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260636
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785744}