chromeos: Treat all unknown-args as the cmd to run in host-side tests.
When running tests that are launched from the host (eg telemetry tests),
the test_runner will take all args passed to it after the last known
arg and treat that as the host-side test command.

Consequently if a known arg is interleaved into the unknown args, the
test runner will not pick it up. By treating unknown args (rather than
trailing args) as the command, we can embed known args into the list.

This will let something like crrev.com/c/2283576 land as-is where we add
'--flash' as the very last arg. Without this CL, that arg is treated
as an arg to telemetry instead of an arg to the test runner.

Bug: 1057152
Change-Id: Id5326dfa3bceab86ae97dc260c9699cf35f56b08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283899
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785787}