CrOS: Network: Fix ShillPropertyHandler for ProhibitedTechnologies
When the event listener was added for shill::kProhibitedTechnologies,
it treated the property as a list (like other properties) however
the property is actually stored as a string in Shill:
https://chromium-review.googlesource.com/c/chromium/src/+/2197136

This was causing error spam when running on linux and in tests.

This fixes that and adds a test.

Bug: 1081488
TBR=tbegin@chromium.org

Change-Id: I502bfe3369d31514b0f6f8e338408d3433ad5cef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283934
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785828}