[ReaderMode] Update appearance controls on page load from stored prefs
https://crrev.com/c/2203775 introduced storing of theme and font family
selections in a user preference. On page load, this was not taken into
account with the radio buttons for the theme selector or the select drop
down for the font family. This change ensures these UI elements are
updated correctly on page load to display the currently active theme
and font family to avoid confusion for users.

This patch also introduces some front end testing for the settings
dialog.

Bug: 1016615, 1070724
Change-Id: I2e45334d012b6d88e813c4bfd4f395adfa309413
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208871
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785829}