[privacy_budget] Use the global settings object for state.
`IdentifiabilityStudyState` was using a separate instance of
`IdentifiabilityStudySettings` to access the externally supplied study
parameters. Instead it now uses the global singleton.

Why do this? There really should be only one settings object. Having the
`IdentifiabilityStudyState` rely on the authoritative instance which is
the same one that the renderer and the rest of the reporting
infrastructure also relies on. In turn, this means that race conditions
involving setting up the global instance are much more visible and are
immediately caught by tests.

Existing tests exercise the functionality being changed here.

R=pauljensen@chromium.org

Bug: 1100424
Bug: 973801
Change-Id: I102495c70faa03378a392aecf9ab6b8c36364ee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275478
Commit-Queue: Asanka Herath <asanka@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785855}