[ozone] Added propagation of drop event modifiers from the platform.
The drag and drop client is supposed to set event flags according to the
current state of modifier keys (i.e., Ctrl, Shift, and Alt).  It was
not implemented in Ozone, but migration of Linux/X11 to Ozone requires
that.  This CL adds the methods that allow the DDDClientOzone to get
these flags from the platform.

The issue has been found by a test that broke at the attempt to use
DesktopDragDropClientOzone in place of DesktopDragDropClientAuraX11.

Bug: 990756
Change-Id: I2657383011a1bf9d716bb5bfa90dbcf010bba48f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279975
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Auto-Submit: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#785872}