Add metric to check pin state on taskbar without checking for unpin verb
This change is to verify that there is no difference in reported metrics
for checking if the app is pinned to the taskbar with or without
the additional check that the shortcut exposes the unpin from taskbar
shell verb. Once this is verified a subsequent change will go in to
remove the shell verb check.

Bug: 1094395
Change-Id: Ibe70e7b2b54ae5bd9f4cdcccb20af7cb93e9f2da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277105
Commit-Queue: Chris Davis <chrdavis@microsoft.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#785881}