Remove make_subset() from PaintImageBuilder
When making a subset of an image, the original PaintImage's dimensions
were set to those of the subset. This caused confusion when accessing
the width() and height() PaintImage() APIs, as it was not always clear
which dimensions were being returned. This change removes all uses of
subrect in PaintImage and forwards the subrect size to where the actual
draw calls are done.

More details about overall PaintImage effort: crbug.com/1023259
Info about the OOPR-Canvas2D project: crbug.com/1018894

Bug: 1031050
Change-Id: Ia57aa5868f671676e3ecc17dd5f0278ec662c366
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258574
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Jonah Chin <jochin@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#785945}